requirements.txt
# Core deep learning
torch>=1.12.0
torchvision>=0.13.0

# Graph neural networks
dgl>=1.0.0
dgllife>=0.3.0

# HuggingFace (MolFormer, ViT, ESM-2, ProtT5)
transformers>=4.30.0
sentencepiece>=0.1.99
tokenizers>=0.13.0

# Scientific computing
numpy>=1.21.0
scipy>=1.7.0
scikit-learn>=1.0.0
pandas>=1.3.0

# Molecular informatics
rdkit>=2022.3.0

# Visualization / progress
tqdm>=4.62.0
prettytable>=3.3.0
Pillow>=9.0.0

# Config management
yacs>=0.1.8
